chore: move repository interface to Persistence/Repositories
