+++ IRasRag.Test/UnitTests/Application/AuthServiceTest.cs
@@ -4,6 +4,7 @@ using IRasRag.Application.Common.Interfaces.Auth;
 using IRasRag.Application.Common.Interfaces.BackgroundJobs;
 using IRasRag.Application.Common.Interfaces.Email;
 using IRasRag.Application.Common.Interfaces.Persistence;
+using IRasRag.Application.Common.Interfaces.Persistence.Repositories;
 using IRasRag.Application.Common.Models;
 using IRasRag.Application.Common.Models.Auth;
 using IRasRag.Application.DTOs;
